requirements.txt
openai
pandas
python-dateutil
tqdm
nbformat